refactor(context): extract configuration resolver
